Enable -Wconversion, fix map size validation, silence vendored hps
catchchallenger_common_flags gains -Wconversion (it already had -Wall -Wextra).
That is the flag that reports implicit narrowing; -Wall/-Wextra say nothing
about it.

general/hps is vendored and must not be modified, and it produced 230 of the
479 warnings on server/cli alone — enough to bury the ones in our own code.
-isystem cannot reach it because our sources include it by relative path, so
the fix stays on our side: general/hps_quiet.hpp wraps the include in a
diagnostic push/ignored/pop (guarded on __GNUC__, MSVC reports C4068 on an
unknown pragma) and the 12 include sites in 11 files now go through it. The
vendor tree is untouched.

Map_loaderMain.cpp parsed the map width/height with stringtouint32() straight
into the uint8_t field, so the truncation happened BEFORE the 1..127 test:
width="257" became 1 and passed as a valid map. The full parsed value is now
range-checked and narrowed afterwards.

The rest is narrowing made explicit where it is provably bounded, each site
read rather than cast away: container sizes into int/uint32_t loop counters and
protocol length fields, and BaseServerLoadSQL's dbSize, which counted bytes but
was declared with the map-id type (uint16_t) while its three neighbours are
uint32_t, so the printed figure wrapped past 64KB.

test/wall_cap.py and test/all.sh: testingserver.py's 30m cap cut its matrix at
17 of 254 results. Measured 2312s end-to-end with a warm ccache, so the cap goes
to 2h — enough for a cold cache without letting a hung run burn an afternoon.

server/cli goes from 479 to 201 warnings. Validated: testingserver 291/291
(gcc+clang, C++11..C++23, hardened/NOXML, and remote mips/geode/armv6-7-8/amd64
LXC nodes), testingprotocolstate 36/36 with valgrind clean, testingmapmanagement
27/27, testingmap2png 2/2.

+++ server/base/BaseServer/BaseServerLoadDatapack.cpp
@@ -197,7 +197,7 @@ void BaseServer::preload_6_sync_the_datapack()
                         #endif
                         #endif
 
-                        hashBase.update(reinterpret_cast<const unsigned char *>(data.data()),data.size());
+                        hashBase.update(reinterpret_cast<const unsigned char *>(data.data()),(unsigned int)data.size());
 
                         /*useful to debug: hashBase.final(reinterpret_cast<unsigned char *>(ProtocolParsingBase::tempBigBufferForOutput));
                         std::cout << str_tolower(binarytoHexa(ProtocolParsingBase::tempBigBufferForOutput,CATCHCHALLENGER_HASH_SIZE)) << "  " << fullPathFileToOpen.c_str() << std::endl;
@@ -345,7 +345,7 @@ void BaseServer::preload_6_sync_the_datapack()
                         }
                         #endif
 
-                        hashMain.update(reinterpret_cast<const unsigned char *>(data.data()),data.size());
+                        hashMain.update(reinterpret_cast<const unsigned char *>(data.data()),(unsigned int)data.size());
                     }
                 }
                 else
@@ -416,7 +416,7 @@ void BaseServer::preload_6_sync_the_datapack()
                     }
                     #endif
 
-                    hashSub.update(reinterpret_cast<const unsigned char *>(data.data()),data.size());
+                    hashSub.update(reinterpret_cast<const unsigned char *>(data.data()),(unsigned int)data.size());
                 }
                 else
                 {
